Moving modules 'files' and 'yaml' to dir 'utils'.
