+++ cibyl/utils/yaml.py
@@ -32,7 +32,7 @@ def parse(file):
     :raises YAMLError: If the file failed to be loaded.
     """
     try:
-        with open(file, 'r') as buffer:
+        with open(file, 'r', encoding='utf8') as buffer:
             return yaml.safe_load(buffer)
     except (OSError, YAMLLoadError) as ex:
         raise YAMLError(f"Failed to parse file: '{file}'") from ex
