+++ requirements.txt
@@ -1,1 +1,1 @@
-PyYAML==6.0
+pyyaml
